Fall back to a conda-supplied R upon encountering a GLIBCXX error. (#21)
See the corresponding commit (2913eec) in the LTLA/basilisk.utils repository
for more details. The idea is, if we detect a GLIBCXX versioning error, we
switch to the internal R fallback to attempt to execute the 'fun' supplied 
to basiliskRun. This requires developers to specify the Python package imports
that might cause problems, so basiliskStart can attempt a load beforehand and
switch to the fallback if problems are detected. It also greatly restricts the
type of code that 'fun' can contain, i.e., only base R or reticulate functions.