Simplify variable persistence across basiliskRun calls.
The storage environment is now explicitly passed into 'fun', which avoids the
hassle of (i) providing a separate function to get it and (ii) trying to
navigate the environment look-up rules. The former is particularly important to
ensure that persistent storage still works when the last resort fallback is
invoked (and thus basilisk itself will not be available in the fallback R).